+++ engine/src/execution_graph/logic_controllers/BatchJoinProcessing.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <tuple>
+
 #include "BatchProcessing.h"
 #include "BlazingColumn.h"
 #include "LogicPrimitives.h"
@@ -40,6 +42,13 @@ struct TableSchema {
 	{}
 };
 
+/* This function takes in the relational algrebra expression and returns:
+- modified relational algrbra expression
+- join condition
+- filter_statement
+- join type */
+std::tuple<std::string, std::string, std::string, std::string> parseExpressionToGetTypeAndCondition(const std::string & expression);
+
 void parseJoinConditionToColumnIndices(const std::string & condition, std::vector<int> & columnIndices);
 
 void split_inequality_join_into_join_and_filter(const std::string & join_statement, std::string & new_join_statement, std::string & filter_statement);
@@ -51,12 +60,6 @@ public:
 		this->query_graph = query_graph;
 		this->input_.add_port("input_a", "input_b");
 
-		join_partition_size_theshold = 400000000; // Threshold for how big to try to make each partition for joining
-		std::map<std::string, std::string> config_options = context->getConfigOptions();
-		auto it = config_options.find("JOIN_PARTITION_SIZE_THRESHOLD");
-		if (it != config_options.end()){
-			join_partition_size_theshold = std::stoull(config_options["JOIN_PARTITION_SIZE_THRESHOLD"]);
-		}
 		this->max_left_ind = -1;
 		this->max_right_ind = -1;
 
@@ -69,6 +72,8 @@ public:
 
 		this->leftArrayCache = 	ral::cache::create_cache_machine(cache_machine_config);
 		this->rightArrayCache = ral::cache::create_cache_machine(cache_machine_config);
+
+		std::tie(this->expression, this->condition, this->filter_statement, this->join_type) = parseExpressionToGetTypeAndCondition(this->expression);		
 	}
 
 	bool can_you_throttle_my_input() {
@@ -78,65 +83,10 @@ public:
 	std::unique_ptr<TableSchema> left_schema{nullptr};
  	std::unique_ptr<TableSchema> right_schema{nullptr};
 
-	std::unique_ptr<ral::frame::BlazingTable> load_set(BatchSequence & input, std::shared_ptr<ral::cache::CacheMachine> cache, bool load_all){
-		std::size_t bytes_loaded = 0;
-		std::vector<std::unique_ptr<ral::frame::BlazingTable>> tables_loaded;
-		if (input.wait_for_next()){
-			tables_loaded.emplace_back(input.next());
-			bytes_loaded += tables_loaded.back()->sizeInBytes();
-		} else {
-			return nullptr;
-		}
-		// NOTE this used to be like this:
-		// while ((input.has_next_now() && bytes_loaded < join_partition_size_theshold) ||
-		//  	(load_all && input.wait_for_next())) {
-		// with the idea that it would just start processing as soon as there was data to process.
-		// This actually does not make it faster, because it makes it so that there are more chunks to do pairwise joins and therefore more join operations
-		// We may want to revisit or rethink this
-
-		while ((input.wait_for_next() && bytes_loaded < join_partition_size_theshold) ||
-					(load_all && input.wait_for_next())) {
-			tables_loaded.emplace_back(std::move(input.next()));
-			bytes_loaded += tables_loaded.back()->sizeInBytes();
-
-			//If the concatenation so far produces a string overflow, we put back the batch because in this case the order of the elements does not matter
-			if(!load_all && ral::utilities::checkIfConcatenatingStringsWillOverflow(tables_loaded)){
-				bytes_loaded -= tables_loaded.back()->sizeInBytes();
-				cache->addToCache(std::move(tables_loaded.back()));
-				tables_loaded.pop_back();				
-				break;
-			}
-		}
-
-		if (tables_loaded.size() == 0){
-			return nullptr;
-		} else if (tables_loaded.size() == 1){
-			return std::move(tables_loaded[0]);
-		} else {
-			std::vector<ral::frame::BlazingTableView> tables_to_concat(tables_loaded.size());
-			for (std::size_t i = 0; i < tables_loaded.size(); i++){
-				tables_to_concat[i] = tables_loaded[i]->toBlazingTableView();
-			}
-
-			// if we were supposed to load all and strings will overflow, lets at least provide a warning
-			if (load_all && ral::utilities::checkIfConcatenatingStringsWillOverflow(tables_to_concat)){
-				logger->warn("{query_id}|{step}|{substep}|{info}",
-								"query_id"_a=(context ? std::to_string(context->getContextToken()) : ""),
-								"step"_a=(context ? std::to_string(context->getQueryStep()) : ""),
-								"substep"_a=(context ? std::to_string(context->getQuerySubstep()) : ""),
-								"info"_a="In PartwiseJoin::load_set Concatenating Strings will overflow strings length");			
-			}
-			return ral::utilities::concatTables(tables_to_concat);
-		}
-	}
-
-
     std::unique_ptr<ral::frame::BlazingTable> load_left_set(){
 
 		this->max_left_ind++;
-		// need to load all the left side, only if doing a FULL OUTER JOIN
-		bool load_all = this->join_type == OUTER_JOIN ?  true : false;
-		auto table = load_set(this->left_sequence, this->input_.get_cache("input_a"), load_all);
+		std::unique_ptr<ral::frame::BlazingTable> table = this->left_sequence.next();
 		if (not left_schema && table != nullptr) {
 			left_schema = std::make_unique<TableSchema>(table->get_schema(),  table->names());
 		}
@@ -148,9 +98,7 @@ public:
 
 	std::unique_ptr<ral::frame::BlazingTable> load_right_set(){
 		this->max_right_ind++;
-		// need to load all the right side, if doing any type of outer join
-		bool load_all = this->join_type != INNER_JOIN ?  true : false;
-		auto table = load_set(this->right_sequence, this->input_.get_cache("input_b"), load_all);
+		std::unique_ptr<ral::frame::BlazingTable> table = this->right_sequence.next();
 		if (not right_schema && table != nullptr) {
 			right_schema = std::make_unique<TableSchema>(table->get_schema(),  table->names());
 		}
@@ -297,19 +245,6 @@ public:
         this->left_sequence = BatchSequence(this->input_.get_cache("input_a"), this, ordered);
 		this->right_sequence = BatchSequence(this->input_.get_cache("input_b"), this, ordered);
 
-		// lets parse part of the expression here, because we need the joinType before we load
-		std::string new_join_statement, filter_statement;
-		StringUtil::findAndReplaceAll(this->expression, "IS NOT DISTINCT FROM", "=");
-		split_inequality_join_into_join_and_filter(this->expression, new_join_statement, filter_statement);
-
-		// Getting the condition and type of join
-		std::string condition = get_named_expression(new_join_statement, "condition");
-		this->join_type = get_named_expression(new_join_statement, "joinType");
-
-		if (condition == "true") {
-			this->join_type = CROSS_JOIN;
-		}
-
 		std::unique_ptr<ral::frame::BlazingTable> left_batch = nullptr;
 		std::unique_ptr<ral::frame::BlazingTable> right_batch = nullptr;
 		bool done = false;
@@ -321,6 +256,11 @@ public:
 			try {
 
 				if (left_batch == nullptr && right_batch == nullptr){ // first load
+					
+					// before we load anything, lets make sure each side has data to process
+					this->left_sequence.wait_for_next();
+					this->right_sequence.wait_for_next();
+					
 					left_batch = load_left_set();
 					right_batch = load_right_set();
 					this->max_left_ind = 0; // we have loaded just once. This is the highest index for now
@@ -328,7 +268,7 @@ public:
 
 					// parsing more of the expression here because we need to have the number of columns of the tables
 					std::vector<int> column_indices;
-					parseJoinConditionToColumnIndices(condition, column_indices);
+					parseJoinConditionToColumnIndices(this->condition, column_indices);
 					for(int i = 0; i < column_indices.size();i++){
 						if(column_indices[i] >= left_batch->num_columns()){
 							this->right_column_indices.push_back(column_indices[i] - left_batch->num_columns());
@@ -478,6 +418,10 @@ public:
 		return kstatus::proceed;
 	}
 
+	std::string get_join_type() {
+		return join_type;
+	}
+
 private:
 	BatchSequence left_sequence, right_sequence;
 
@@ -486,10 +430,11 @@ private:
 	std::vector<std::vector<bool>> completion_matrix;
 	std::shared_ptr<ral::cache::CacheMachine> leftArrayCache;
 	std::shared_ptr<ral::cache::CacheMachine> rightArrayCache;
-	std::size_t join_partition_size_theshold;
-
+	
 	// parsed expression related parameters
 	std::string join_type;
+	std::string condition;
+	std::string filter_statement;
 	std::vector<cudf::size_type> left_column_indices, right_column_indices;
 	std::vector<cudf::data_type> join_column_common_types;
 	bool normalize_left, normalize_right;
@@ -504,6 +449,8 @@ public:
 		this->query_graph = query_graph;
 		this->input_.add_port("input_a", "input_b");
 		this->output_.add_port("output_a", "output_b");
+
+		std::tie(this->expression, this->condition, this->filter_statement, this->join_type) = parseExpressionToGetTypeAndCondition(this->expression);
 	}
 
 	bool can_you_throttle_my_input() {
@@ -786,17 +733,27 @@ public:
 		bool any_unknowns_left = std::any_of(nodes_num_bytes_left.begin(), nodes_num_bytes_left.end(), [](int64_t bytes){return bytes == -1;});
 		bool any_unknowns_right = std::any_of(nodes_num_bytes_right.begin(), nodes_num_bytes_right.end(), [](int64_t bytes){return bytes == -1;});
 
-		if (any_unknowns_left || any_unknowns_right){
-			return std::make_pair(false, false); // we wont do any small table scatter if we have unknowns
-		}
-
 		int64_t total_bytes_left = std::accumulate(nodes_num_bytes_left.begin(), nodes_num_bytes_left.end(), int64_t(0));
 		int64_t total_bytes_right = std::accumulate(nodes_num_bytes_right.begin(), nodes_num_bytes_right.end(), int64_t(0));
 
-		int num_nodes = context->getTotalNodes();
-
 		bool scatter_left = false;
 		bool scatter_right = false;
+		if (any_unknowns_left || any_unknowns_right){
+			// with CROSS_JOIN we want to scatter or or the other, no matter what, even with unknowns
+			if (this->join_type == CROSS_JOIN){
+				if(total_bytes_left < total_bytes_right) {
+					scatter_left = true;
+				} else {
+					scatter_right = true;
+				}
+				return std::make_pair(scatter_left, scatter_right);
+			} else {
+				return std::make_pair(false, false); // we wont do any small table scatter if we have unknowns
+			}
+		}
+
+		int num_nodes = context->getTotalNodes();
+
 		int64_t estimate_regular_distribution = (total_bytes_left + total_bytes_right) * (num_nodes - 1) / num_nodes;
 		int64_t estimate_scatter_left = (total_bytes_left) * (num_nodes - 1);
 		int64_t estimate_scatter_right = (total_bytes_right) * (num_nodes - 1);
@@ -808,14 +765,29 @@ public:
 			max_join_scatter_mem_overhead = std::stoull(config_options["MAX_JOIN_SCATTER_MEM_OVERHEAD"]);
 		}
 
-		if(estimate_scatter_left < estimate_regular_distribution ||
-			estimate_scatter_right < estimate_regular_distribution) {
-			if(estimate_scatter_left < estimate_scatter_right &&
-				total_bytes_left < max_join_scatter_mem_overhead) {
+		// with CROSS_JOIN we want to scatter or or the other
+		if (this->join_type == CROSS_JOIN){
+			if(estimate_scatter_left < estimate_scatter_right) {
 				scatter_left = true;
-			} else if(estimate_scatter_right < estimate_scatter_left &&
-					total_bytes_right < max_join_scatter_mem_overhead) {
+			} else {
 				scatter_right = true;
+			}
+		// with LEFT_JOIN we cant scatter the left side
+		} else if (this->join_type == LEFT_JOIN) {
+			if(estimate_scatter_right < estimate_regular_distribution &&
+						total_bytes_right < max_join_scatter_mem_overhead) {
+				scatter_right = true;
+			}
+		} else {
+			if(estimate_scatter_left < estimate_regular_distribution ||
+				estimate_scatter_right < estimate_regular_distribution) {
+				if(estimate_scatter_left < estimate_scatter_right &&
+					total_bytes_left < max_join_scatter_mem_overhead) {
+					scatter_left = true;
+				} else if(estimate_scatter_right < estimate_scatter_left &&
+						total_bytes_right < max_join_scatter_mem_overhead) {
+					scatter_right = true;
+				}
 			}
 		}
 		return std::make_pair(scatter_left, scatter_right);
@@ -1039,13 +1011,6 @@ public:
 		BatchSequence left_sequence(this->input_.get_cache("input_a"), this, ordered);
 		BatchSequence right_sequence(this->input_.get_cache("input_b"), this, ordered);
 
-		// lets parse part of the expression here, because we need the joinType before we load
-		std::string new_join_statement, filter_statement;
-		StringUtil::findAndReplaceAll(this->expression, "IS NOT DISTINCT FROM", "=");
-		split_inequality_join_into_join_and_filter(this->expression, new_join_statement, filter_statement);
-		std::string condition = get_named_expression(new_join_statement, "condition");
-		this->join_type = get_named_expression(new_join_statement, "joinType");
-
 		std::unique_ptr<ral::frame::BlazingTable> left_batch = left_sequence.next();
 		std::unique_ptr<ral::frame::BlazingTable> right_batch = right_sequence.next();
 
@@ -1067,9 +1032,6 @@ public:
 		} else {
 			scatter_left_right = determine_if_we_are_scattering_a_small_table(left_batch->toBlazingTableView(),
 																				right_batch->toBlazingTableView());
-			if (scatter_left_right.first && this->join_type == LEFT_JOIN){
-				scatter_left_right.first = false; // cant scatter the left side for a left outer join
-			}
 		}
 		// scatter_left_right = std::make_pair(false, false); // Do this for debugging if you want to disable small table join optmization
 		if (scatter_left_right.first){
@@ -1120,9 +1082,15 @@ public:
 		return kstatus::proceed;
 	}
 
+	std::string get_join_type() {
+		return join_type;
+	}
+
 private:
 	// parsed expression related parameters
 	std::string join_type;
+	std::string condition;
+	std::string filter_statement;
 	std::vector<cudf::size_type> left_column_indices, right_column_indices;
 	std::vector<cudf::data_type> join_column_common_types;
 	bool normalize_left, normalize_right;