+++ engine/src/execution_graph/logic_controllers/CacheMachine.cpp
@@ -28,13 +28,17 @@ std::string randomString(std::size_t length) {
 	return random_string;
 }
 
-size_t CacheDataLocalFile::sizeInBytes() const {
+size_t CacheDataLocalFile::fileSizeInBytes() const {
 	struct stat st;
 
 	if(stat(this->filePath_.c_str(), &st) == 0)
 		return (st.st_size);
 	else
 		throw;
+}
+
+size_t CacheDataLocalFile::sizeInBytes() const {
+	return size_in_bytes;
 }
 
 std::unique_ptr<ral::frame::BlazingTable> CacheDataLocalFile::decache() {
@@ -50,9 +54,9 @@ std::unique_ptr<ral::frame::BlazingTable> CacheDataLocalFile::decache() {
 CacheDataLocalFile::CacheDataLocalFile(std::unique_ptr<ral::frame::BlazingTable> table, std::string orc_files_path)
 	: CacheData(CacheDataType::LOCAL_FILE, table->names(), table->get_schema(), table->num_rows())
 {
+	this->size_in_bytes = table->sizeInBytes();
 	this->filePath_ = orc_files_path + "/.blazing-temp-" + randomString(64) + ".orc";
 
-	std::cout << "CacheDataLocalFile: " << this->filePath_ << std::endl;
 	cudf_io::table_metadata metadata;
 	for(auto name : table->names()) {
 		metadata.column_names.emplace_back(name);
@@ -76,9 +80,7 @@ CacheMachine::CacheMachine(std::shared_ptr<Context> context): ctx(context), cach
 	this->memory_resources.push_back( &blazing_disk_memory_resource::getInstance() );
 	this->num_bytes_added = 0;
 	this->num_rows_added = 0;
-	this->flow_control_batches_threshold = std::numeric_limits<std::uint32_t>::max();
 	this->flow_control_bytes_threshold = std::numeric_limits<std::size_t>::max();
-	this->flow_control_batches_count = 0;
 	this->flow_control_bytes_count = 0;
 
 	logger = spdlog::get("batch_logger");
@@ -95,7 +97,7 @@ CacheMachine::CacheMachine(std::shared_ptr<Context> context): ctx(context), cach
 							"kernel_type"_a="cache");
 }
 
-CacheMachine::CacheMachine(std::shared_ptr<Context> context, std::uint32_t flow_control_batches_threshold, std::size_t flow_control_bytes_threshold) : ctx(context), cache_id(CacheMachine::cache_count)
+CacheMachine::CacheMachine(std::shared_ptr<Context> context, std::size_t flow_control_bytes_threshold) : ctx(context), cache_id(CacheMachine::cache_count)
 {
 	CacheMachine::cache_count++;
 
@@ -105,9 +107,7 @@ CacheMachine::CacheMachine(std::shared_ptr<Context> context, std::uint32_t flow_
 	this->memory_resources.push_back( &blazing_disk_memory_resource::getInstance() );
 	this->num_bytes_added = 0;
 	this->num_rows_added = 0;
-	this->flow_control_batches_threshold = flow_control_batches_threshold;
 	this->flow_control_bytes_threshold = flow_control_bytes_threshold;
-	this->flow_control_batches_count = 0;
 	this->flow_control_bytes_count = 0;
 
 	logger = spdlog::get("batch_logger");
@@ -143,7 +143,6 @@ void CacheMachine::finish() {
 									"info"_a="CacheMachine finish()",
 									"duration"_a="",
 									"cache_id"_a=cache_id);
-	
 }
 
 bool CacheMachine::is_finished() {
@@ -172,7 +171,6 @@ void CacheMachine::addHostFrameToCache(std::unique_ptr<ral::frame::BlazingHostTa
 									"rows"_a=host_table->num_rows());
 
 		std::unique_lock<std::mutex> lock(flow_control_mutex);
-		flow_control_batches_count++;
 		flow_control_bytes_count += host_table->sizeInBytes();
 		lock.unlock();
 
@@ -203,7 +201,6 @@ void CacheMachine::addCacheData(std::unique_ptr<ral::cache::CacheData> cache_dat
 	// we dont want to add empty tables to a cache, unless we have never added anything
 	if ((!this->something_added || cache_data->num_rows() > 0) || always_add){
 		std::unique_lock<std::mutex> lock(flow_control_mutex);
-		flow_control_batches_count++;
 		flow_control_bytes_count += cache_data->sizeInBytes();
 		lock.unlock();
 
@@ -264,32 +261,16 @@ void CacheMachine::addCacheData(std::unique_ptr<ral::cache::CacheData> cache_dat
 }
 
 void CacheMachine::addToCache(std::unique_ptr<ral::frame::BlazingTable> table, const std::string & message_id, bool always_add) {
-
 	// we dont want to add empty tables to a cache, unless we have never added anything
 	if (!this->something_added || table->num_rows() > 0|| always_add){
 		for (auto col_ind = 0; col_ind < table->num_columns(); col_ind++){
 			if (table->view().column(col_ind).offset() > 0){
 				table->ensureOwnership();
-			}
-
-			if (table->view().column(col_ind).offset() > 0){
-				std::string err = "ERROR: Add to CacheMachine into cache table column " + table->names()[col_ind] + " has offset";
-				logger->error("{query_id}|{step}|{substep}|{info}|{duration}|kernel_id|{kernel_id}|offset|{offset}",
-								"query_id"_a=(ctx ? std::to_string(ctx->getContextToken()) : ""),
-								"step"_a=(ctx ? std::to_string(ctx->getQueryStep()) : ""),
-								"substep"_a=(ctx ? std::to_string(ctx->getQuerySubstep()) : ""),
-								"info"_a=err,
-								"duration"_a="",
-								"kernel_id"_a=message_id,
-								"offset"_a=table->view().column(col_ind).offset());
-				throw err;
+				break;
 			}
 
 		}
-
-		
 		std::unique_lock<std::mutex> lock(flow_control_mutex);
-		flow_control_batches_count++;
 		flow_control_bytes_count += table->sizeInBytes();
 		lock.unlock();
 
@@ -310,12 +291,9 @@ void CacheMachine::addToCache(std::unique_ptr<ral::frame::BlazingTable> table, c
 						"rows"_a=table->num_rows());
 
 					// before we put into a cache, we need to make sure we fully own the table
-					auto column_names = table->names();
-					auto cudf_table = table->releaseCudfTable();
-					std::unique_ptr<ral::frame::BlazingTable> fully_owned_table =
-						std::make_unique<ral::frame::BlazingTable>(std::move(cudf_table), column_names);
-
-					auto cache_data = std::make_unique<GPUCacheData>(std::move(fully_owned_table));
+					table->ensureOwnership();
+					
+					auto cache_data = std::make_unique<GPUCacheData>(std::move(table));
 					auto item =	std::make_unique<message>(std::move(cache_data), message_id);
 					this->waitingCache->put(std::move(item));
 
@@ -379,7 +357,6 @@ std::unique_ptr<ral::frame::BlazingTable> CacheMachine::get_or_wait(size_t index
 
 	std::unique_ptr<ral::frame::BlazingTable> output = message_data->get_data().decache();
 	std::unique_lock<std::mutex> lock(flow_control_mutex);
-	flow_control_batches_count--;
 	flow_control_bytes_count -= output->sizeInBytes();
 	flow_control_condition_variable.notify_all();
 	return std::move(output);
@@ -402,7 +379,6 @@ std::unique_ptr<ral::frame::BlazingTable> CacheMachine::pullFromCache() {
 
 	std::unique_ptr<ral::frame::BlazingTable> output = message_data->get_data().decache();
 	std::unique_lock<std::mutex> lock(flow_control_mutex);
-	flow_control_batches_count--;
 	flow_control_bytes_count -= output->sizeInBytes();
 	flow_control_condition_variable.notify_all();
 	return std::move(output);
@@ -425,12 +401,10 @@ std::unique_ptr<ral::cache::CacheData> CacheMachine::pullCacheData(std::string m
 								"rows"_a=message_data->get_data().num_rows());
 									std::unique_ptr<ral::cache::CacheData> output = message_data->release_data();
 	std::unique_lock<std::mutex> lock(flow_control_mutex);
-	flow_control_batches_count--;
 	flow_control_bytes_count -= output->sizeInBytes();
 	flow_control_condition_variable.notify_all();
 	return std::move(output);
 }
-
 
 std::unique_ptr<ral::frame::BlazingTable> CacheMachine::pullUnorderedFromCache() {
 
@@ -462,7 +436,6 @@ std::unique_ptr<ral::frame::BlazingTable> CacheMachine::pullUnorderedFromCache()
 
 		std::unique_ptr<ral::frame::BlazingTable> output = message_data->get_data().decache();
 		std::unique_lock<std::mutex> lock(flow_control_mutex);
-		flow_control_batches_count--;
 		flow_control_bytes_count -= output->sizeInBytes();
 		flow_control_condition_variable.notify_all();
 		return std::move(output);
@@ -488,16 +461,15 @@ std::unique_ptr<ral::cache::CacheData> CacheMachine::pullCacheData() {
 
 	std::unique_ptr<ral::cache::CacheData> output = message_data->release_data();
 	std::unique_lock<std::mutex> lock(flow_control_mutex);
-	flow_control_batches_count--;
 	flow_control_bytes_count -= output->sizeInBytes();
 	flow_control_condition_variable.notify_all();
 	return std::move(output);
 }
 
 
-bool CacheMachine::thresholds_are_met(std::uint32_t batches_count, std::size_t bytes_count){
+bool CacheMachine::thresholds_are_met(std::size_t bytes_count){
 
-	return batches_count > this->flow_control_batches_threshold && bytes_count > this->flow_control_bytes_threshold;
+	return bytes_count > this->flow_control_bytes_threshold;
 }
 
 void CacheMachine::wait_if_cache_is_saturated() {
@@ -506,7 +478,7 @@ void CacheMachine::wait_if_cache_is_saturated() {
 
 	std::unique_lock<std::mutex> lock(flow_control_mutex);
 	while(!flow_control_condition_variable.wait_for(lock, 60000ms, [&, this] {
-			bool cache_not_saturated = !thresholds_are_met(flow_control_batches_count, flow_control_bytes_count);
+			bool cache_not_saturated = !thresholds_are_met(flow_control_bytes_count);
 
 			if (!cache_not_saturated && blazing_timer.elapsed_time() > 59000){
 				logger->warn("{query_id}|{step}|{substep}|{info}|{duration}||||",
@@ -584,34 +556,40 @@ size_t CacheMachine::downgradeCacheData() {
 ConcatenatingCacheMachine::ConcatenatingCacheMachine(std::shared_ptr<Context> context)
 	: CacheMachine(context) {}
 
-ConcatenatingCacheMachine::ConcatenatingCacheMachine(std::shared_ptr<Context> context, std::uint32_t flow_control_batches_threshold, std::size_t flow_control_bytes_threshold, bool concat_all)
-	: CacheMachine(context, flow_control_batches_threshold, flow_control_bytes_threshold), concat_all(concat_all) {}
+ConcatenatingCacheMachine::ConcatenatingCacheMachine(std::shared_ptr<Context> context, std::size_t flow_control_bytes_threshold, bool concat_all)
+	: CacheMachine(context, flow_control_bytes_threshold), concat_all(concat_all) {}
 
 // This method does not guarantee the relative order of the messages to be preserved
 std::unique_ptr<ral::frame::BlazingTable> ConcatenatingCacheMachine::pullFromCache() {
+
+	if (concat_all){
+		waitingCache->wait_until_finished();
+	} else {
+		waitingCache->wait_until_num_bytes(this->flow_control_bytes_threshold);
+	}
 
 	size_t total_bytes = 0;
 	std::vector<std::unique_ptr<message>> collected_messages;
 	std::unique_ptr<message> message_data;
 	std::string message_id = "";
-	while (message_data = waitingCache->pop_or_wait())
-	{
-		auto& cache_data = message_data->get_data();
-		if (concat_all || collected_messages.empty() || !thresholds_are_met(1 + collected_messages.size(), total_bytes + cache_data.sizeInBytes())) {
-			total_bytes += cache_data.sizeInBytes();
-			message_id = message_data->get_message_id();
-			collected_messages.push_back(std::move(message_data));
 
-			// we need to decrement here and not at the end, otherwise we can end up with a dead lock
-			std::unique_lock<std::mutex> lock(flow_control_mutex);
-			flow_control_batches_count--;
-			flow_control_bytes_count -= cache_data.sizeInBytes();
-			flow_control_condition_variable.notify_all();
-		} else {
-			waitingCache->put(std::move(message_data));
+	do {
+		message_data = waitingCache->pop_or_wait();
+		if (message_data == nullptr){
 			break;
 		}
-	}
+		auto& cache_data = message_data->get_data();
+		total_bytes += cache_data.sizeInBytes();
+		message_id = message_data->get_message_id();
+		collected_messages.push_back(std::move(message_data));
+
+		// we need to decrement here and not at the end, otherwise we can end up with a dead lock
+		std::unique_lock<std::mutex> lock(flow_control_mutex);
+		flow_control_bytes_count -= cache_data.sizeInBytes();
+		flow_control_condition_variable.notify_all();
+
+	} while (!thresholds_are_met(total_bytes + waitingCache->get_next_size_in_bytes()));
+
 	std::unique_ptr<ral::frame::BlazingTable> output;
 	size_t num_rows = 0;
 	if(collected_messages.empty()){
@@ -642,7 +620,6 @@ std::unique_ptr<ral::frame::BlazingTable> ConcatenatingCacheMachine::pullFromCac
 				collected_messages[i] =	std::make_unique<message>(std::move(cache_data), collected_messages[i]->get_message_id());
 				std::unique_lock<std::mutex> lock(flow_control_mutex);
 				for (; i < collected_messages.size(); i++){
-					flow_control_batches_count++;
 					flow_control_bytes_count += collected_messages[i]->get_data().sizeInBytes();
 					this->waitingCache->put(std::move(collected_messages[i]));
 				}