+++ engine/src/execution_graph/logic_controllers/LogicPrimitives.h
@@ -59,6 +59,7 @@ public:
 
 	unsigned long long sizeInBytes();
 	void ensureOwnership();
+
 private:
 	std::vector<std::string> columnNames;
 	std::vector<std::unique_ptr<BlazingColumn>> columns;