+++ engine/src/utilities/DebuggingUtils.h
@@ -2,7 +2,7 @@
 #define _BLAZING_DEBUG_UTILS_H
 
 #include "execution_graph/logic_controllers/LogicPrimitives.h"
-#include "execution_graph/logic_controllers/CacheMachine.h"
+// #include "execution_graph/logic_controllers/CacheMachine.h"
 
 namespace ral {
 namespace utilities {
@@ -15,7 +15,7 @@ void print_blazing_table_view_schema(ral::frame::BlazingTableView table_view, co
 
 std::string blazing_table_view_schema_to_string(ral::frame::BlazingTableView table_view, const std::string table_name);
 
-std::string cache_data_schema_to_string(ral::cache::CacheData * cache_data);
+// std::string cache_data_schema_to_string(ral::cache::CacheData * cache_data);
 
 }  // namespace utilities
 }  // namespace ral